# backend/tests/snapshots/baseline.openapi.json
{
  "components": {
    "schemas": {
      "Body_create_profile_route_api_restyle_profile_post": {
        "properties": {
          "selfie": {
            "contentMediaType": "application/octet-stream",
            "title": "Selfie",
            "type": "string"
          }
        },
        "required": [
          "selfie"
        ],
        "title": "Body_create_profile_route_api_restyle_profile_post",
        "type": "object"
      },
...
      "Body_start_restyle_api_restyle_post": {
        "properties": {
          "background_prompt": {
            "title": "Background Prompt",
            "type": "string"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "lighting_prompt": {
            "title": "Lighting Prompt",
            "type": "string"
          }
        },
        "required": [
          "file",
          "background_prompt",
          "lighting_prompt"
        ],
        "title": "Body_start_restyle_api_restyle_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RestyleStatus": {
        "properties": {
          "logs": {
            "items": {
              "type": "string"
            },
            "title": "Logs",
            "type": "array"
          },
          "progress_pct": {
            "default": 0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress Pct",
            "type": "integer"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "logs"
        ],
        "title": "RestyleStatus",
        "type": "object"
      },
...
      "SelectRequest": {
        "properties": {
          "idx": {
            "maximum": 99.0,
            "minimum": 1.0,
            "title": "Idx",
            "type": "integer"
          }
        },
        "required": [
          "idx"
        ],
        "title": "SelectRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/restyle": {
      "post": {
        "description": "Start a restyle job. Returns ``{job_id}`` immediately; poll\n``GET /api/restyle/{job_id}`` for status.",
        "operationId": "start_restyle_api_restyle_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_start_restyle_api_restyle_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Restyle"
      }
    },
    "/api/restyle/profile": {
      "post": {
        "description": "Create profile + kick off async 5-background generation. Returns {profile_id}.",
        "operationId": "create_profile_route_api_restyle_profile_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_profile_route_api_restyle_profile_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Profile Route"
      }
    },
    "/api/restyle/profile/{profile_id}": {
      "get": {
        "description": "Return current profile state (status, generated count, selected idx, bg URLs).",
        "operationId": "get_profile_route_api_restyle_profile__profile_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Profile Route"
      }
    },
    "/api/restyle/profile/{profile_id}/regenerate": {
      "post": {
        "description": "Re-run the 5-background generation against the saved selfie.",
        "operationId": "regenerate_route_api_restyle_profile__profile_id__regenerate_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate Route"
      }
    },
    "/api/restyle/profile/{profile_id}/select": {
      "post": {
        "description": "Mark one of the generated backgrounds as active. 400 if idx out of range.",
        "operationId": "select_background_route_api_restyle_profile__profile_id__select_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Background Route"
      }
    },
    "/api/restyle/{job_id}": {
      "get": {
        "description": "Poll the status of a restyle job.",
        "operationId": "restyle_status_api_restyle__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestyleStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restyle Status"
      }
    },
...
    "/profiles/{profile_id}/{filename}": {
      "get": {
        "description": "Static-serve selfie.png + bg-N.png with allowlist + path traversal guard.\n\nNo X-Gemini-Key required: the profile_id UUID (128-bit random) is the\ncapability. Consistent with GET /api/restyle/{job_id} (also unauthenticated;\njob_id is the credential).",
        "operationId": "serve_profile_file_profiles__profile_id___filename__get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Serve Profile File"
      }
    },
...
  }
}